refactor: component props and types for improved type organization
