+++ src/components/dashboard/StatItem.tsx
@@ -1,9 +1,6 @@
-interface StatItemProps {
-  title: string;
-  data: string;
-}
+import { CardProps } from '@/types/Components.types';
 
-const StatItem = ({ title, data }: StatItemProps) => (
+const StatItem = ({ title, data }: CardProps) => (
   <div className="flex flex-col gap-1.5 justify-between items-center py-6 lg:px-8 border-b-[1px] last:border-b-0">
     <span className="text-secondary-foreground text-[16px] font-semibold tracking-[0.3px] text-center">
       {title}
