+++ src/types/Layout.types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface LayoutProps {
+  children: ReactNode;
+}
